Levels now work properly
The color problem is solved. There is now a Map in a Map to differentiate obsels from the same button but not from the same state.
The functions getSameObselsColor() and updateObselsColor() have been rewritten to match the new architecture
The state (ie, the one which an obsel come from) of the obsel is added as a class on the DOM element (like the group)