It works !!
The function added to the fsm prototype does a little trick to the state of the obsel. In level's description, the "states" object tells for each button which states to regroup (it works like an associative array).

Changed object "states" in level's description
Changed levels ID to be unique
Modified level 2 (which was actually level 3)
Added level 3 (because it was almost finished)