Level 4 created
It took me 10-15min to draw the state machine and approximatively 10min to write the file. So this JSON description seems to be pretty fast and useful (for now).
(The level isn't tested yet, as well as the level 3, because of Wamp... If only I can know why...)

Created level 4
Added level 3 and 4 to levels menu description