Save user level
User level (ie, the name of the level, the trace, the score and wether it's finished or not) is saved in a object. This object isn't permanently saved for now, it only works when the level is being played.
The export button in the menu only print the user save object in the console at the moment

Added an export button in the menu
Modified variable userSave
Modified loadLevel and resetPlayground accordingly
Added exportSave() function (only print in the console)
Modified addObsel, updateScore and winLevel to work correctly with the user save object