Export save
Creating and saving a file using JavaScript is a real mess (to prevent security issues) so in order to save the current level, the JSON file is prompted to the user who will have to copy and paste it in a JSON file.
I will search another (more practical) way to do it.

Changed exportSave function to prompt JSON object of the save
Modified i18n files for the instructions