Easier save loading
The player do not need to load the level (s)he wants to import a save to, it is done automatically after the save's file verification.
Be careful, to achieve that the ajax call to load the specified level's playground is synchronous (despite all warnings about the subject), so a refactor is needed when a better solution will be found.

Changed levels unique ID (for easier manipulation)
Changed ajax function, a third parameter (async) can be sent
Changed onclick listener on levelLink in the menu to use the new ajax call
Changed importSave function, instead of testing the current level, load the corresponding level before importing