Added the Three.js scene
It is a simple scene with a cube over an horizontal plane with the possibility to change the angle/position/zoom of the camera.
It doesn't work as I wanted (the scene is created on the opening of panel), so it creates some problems when the panel is opened several times.

Added three.js and OrbitControls (sort of plugin of three.js)
Added the creation of a simple scene