Tried to separate WebGL initialization and scene launching
I tried to keep the same three.js scene for all level's 3D representation but I don't manage to clear correctly the scene.
I also changed the way the first level is loaded. There's still a problem the first time because loading the level 0 launch the onhashchange which load the level 0 a second time...